+++ Runtime/ForceReloadAttribute.cs
@@ -1,9 +1,0 @@
-using System;
-using System.Diagnostics;
-
-namespace EasyTypeReload
-{
-    [Conditional("UNITY_EDITOR")]
-    [AttributeUsage(AttributeTargets.Field, AllowMultiple = false)]
-    public sealed class ForceReloadAttribute : Attribute { }
-}