+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `PreserveReadonly` property in `ReloadOnEnterPlayModeAttribute`.
+
+### Changed
+
+- Package name.
+- Package description.
+
+### Removed
+
+- `ForceReloadAttribute`.
+
 ## [1.1.0] - 2024-01-26
 
 ### Added